Move saveModeledMethods to a separate file
